Improve server generatation and server data handling, seperate ollama client from other code.

+++ server_provisioner.py
@@ -1,12 +1,15 @@
 from logging_config import logger
 import time
 
-from model.server import Server
+from model.hetzner_cloud_server_instance import HetznerCloudServerInstance
+from ollama.ollama_client import OllamaClient
+from ollama.ollama_test_run import OllamaTestRun
 
 # Konstanten
 RETRY_DELAY = 20  # Sekunden
-START_DELAY = 20 # Sekunden
-RESTART_DELAY = 20 # Sekunden
+START_DELAY = 20  # Sekunden
+RESTART_DELAY = 20  # Sekunden
+
 
 class ServerProvisioner:
     def __init__(self, config, hcloud_manager, notifier, ssh_manager, ollama_manager):
@@ -30,95 +33,56 @@ class ServerProvisioner:
             logger.error(f"Server status check failed: {exception}")
         return False
 
-    def manage_server(self, serverModel: Server):
-
+    def manage_server(self, server_model: HetznerCloudServerInstance):
         try:
-            self.ssh_manager.set_hostname(serverModel.ipv4)
+            self.ssh_manager.set_hostname(server_model.get_ipv4())
             if not self.ssh_manager.connect():
-                logger.error(f"Server Provisioning Failed. SSH connection to server '{serverModel.name}' failed.")
+                logger.error(
+                    f"Server Provisioning Failed. SSH connection to server '{server_model.get_server_name()}' failed.")
                 self.notifier.send_email(
                     subject="Server Provisioning Failed",
-                    message=f"SSH connection to server '{serverModel.name}' failed."
+                    message=f"SSH connection to server '{server_model.get_server_name()}' failed."
                 )
                 return
 
-            self.ollama_manager.install_ollama()
+            self.ollama_manager.install_ollama(server_model)
 
         finally:
             self.ssh_manager.close()
 
-    def manage_ollama(self, serverModel: Server):
-        if not self.ollama_manager.is_ollama_ready():
-            logger.error(f"Ollama Installation Failed. Ollama on server '{serverModel.name}' is not responding.")
-            self.notifier.send_email(
-                subject="Ollama Installation Failed",
-                message=f"Ollama on server '{serverModel.name}' is not responding."
-            )
-            return
+    def manage_ollama(self, server_model: HetznerCloudServerInstance):
+        ollama_client = OllamaClient(self.config, server_model.get_ipv4())
 
-        # Load Model
-        model_name = "deepseek-r1:1.5b"
-        result = self.ollama_manager.download_model(model_name)
-        if result:
-            logger.debug(f"Model '{model_name}' download initiated successfully.")
-        else:
-            logger.error(f"Failed to initiate download for model '{model_name}'.")
-
-        # Start Chat
-        user_message = "Hello, how are you?"
-        logger.info(f"Model's request: {user_message}")
-        reply = self.ollama_manager.send_chat_message(model_name, user_message)
-        logger.info(f"Model's response: {reply}")
-
-        user_message = "Please tell me a story about LLM fight humans in 300 words."
-        logger.info(f"Model's request: {user_message}")
-        reply = self.ollama_manager.send_chat_message(model_name, user_message)
-        logger.info(f"Model's response: {reply}")
-
-        user_message = "Thanks for the test. Bye Bye"
-        logger.info(f"Model's request: {user_message}")
-        reply = self.ollama_manager.send_chat_message(model_name, user_message)
-        logger.info(f"Model's response: {reply}")
+        ollama_test_run = OllamaTestRun(self.config, ollama_client, self.notifier)
+        ollama_test_run.run_test()
 
     def create_server(self):
-        hetzner_ssh_key_name = self.config.get('hetzner','ssh_key_name')
+        hetzner_ssh_key_name = self.config.get('hetzner', 'ssh_key_name')
         server_name = "ollama-server"
         server_type = "cax21"  # Beispiel für einen ARM-Servertyp
         image_name = "ubuntu-22.04"
         server_location = "nbg1"
 
-        server = self.hcloud_manager.create_server(server_name, server_type, image_name, server_location, hetzner_ssh_key_name)
+        server = self.hcloud_manager.create_server(server_name, server_type, image_name, server_location,
+                                                   hetzner_ssh_key_name)
 
         time.sleep(START_DELAY)
 
         return server
 
+    def process_server(self):
+        server_instance_id = self.config.get('hetzner', 'server_instance_id', fallback=None)
+
+        logger.debug(f"Server instance_id: {server_instance_id}")
+
+        if server_instance_id:
+            logger.debug("Use existing server instance.")
+            self.use_server(server_instance_id)
+        else:
+            logger.debug("Provisioning new server instance.")
+            self.provision_server()
+
     def provision_server(self):
-
-        # try:
-        #     server = self.hcloud_manager.get_server(60517182)
-        # except Exception as ex:
-        #      logger.error(f"Get server state failed: {ex}")
-        #
-        # logger.debug("Server Id: " + str(server.id))
-        # logger.debug("Server Name: " + str(server.name))
-        # logger.debug("Server Status: " + str(server.status))
-        # logger.debug("Server IPv4: " + str(server.public_net.ipv4.ip))
-        # logger.debug("Server IPv6: " + str(server.public_net.ipv6.ip))
-        #
-        # serverModel = Server(server.id, server.name, server.status, server.public_net.ipv4.ip,
-        #                       server.public_net.ipv6.ip)
-        # self.ssh_manager.set_hostname(serverModel.ipv4)
-        #
-        # self.manage_server(serverModel)
-        #
-        # self.hcloud_manager.reboot_server(server)
-        # time.sleep(RESTART_DELAY)
-        #
-        # if self.check_server_status(server):
-        #     self.manage_ollama(serverModel)
-
-        #return
 
         server = self.create_server()
 
@@ -129,8 +93,8 @@ class ServerProvisioner:
                 except Exception as ex:
                     logger.error(f"Get server state failed: {ex}")
 
-                serverModel = Server(server.id, server.name, server.status, server.public_net.ipv4.ip,
-                                     server.public_net.ipv6.ip)
+                serverModel = HetznerCloudServerInstance(server)
+                logger.debug(serverModel)
 
                 self.manage_server(serverModel)
 
@@ -140,8 +104,23 @@ class ServerProvisioner:
                 if self.check_server_status(server):
                     self.manage_ollama(serverModel)
 
-        if not self.hcloud_manager.delete_server(server):
-           self.notifier.send_email(
-               subject="Server Deletion Failed",
-               message=f"Failed to delete server '{server.name}' after multiple attempts."
-           )
+        delete = self.config.get('hetzner', 'delete', fallback=True).lower().strip()
+        if delete == "true":
+            if not self.hcloud_manager.delete_server(server):
+                self.notifier.send_email(
+                    subject="Server Deletion Failed",
+                    message=f"Failed to delete server '{server.name}' after multiple attempts."
+                )
+
+    def use_server(self, server_instance_id):
+        try:
+            server = self.hcloud_manager.get_server(server_instance_id)
+
+            serverModel = HetznerCloudServerInstance(server)
+
+            logger.debug(serverModel)
+
+            if self.check_server_status(server):
+                self.manage_ollama(serverModel)
+        except Exception as ex:
+            logger.error(f"Get use server failed: {ex}")
